Keep linux mounts for linux sandboxes on Windows/Darwin
